Correction de bugs pour les points.

+++ include/Hud.hpp
@@ -7,9 +7,9 @@
 class Hud
 {
     public:
-        Hud(unsigned short p_life, unsigned long p_score, sf::Font p_font, sf::Vector2f p_tailleFenetre);
+        Hud(unsigned short p_life, unsigned int p_score, sf::Font p_font, sf::Vector2f p_tailleFenetre);
         void updateLife(unsigned short n_life);
-        void updateScore(unsigned long n_score);
+        void updateScore(unsigned int n_score);
         void setSizeWindow(sf::Vector2f n_sizeWindow);
         sf::Text getLife();
         sf::Text getScore();
